pyproject.toml
# 构建系统相关配置
[build-system]
requires = ["setuptools >= 64", "wheel"]
build-backend = "setuptools.build_meta"
[tool.pytest.ini_options]
pythonpath = ["."]
# 项目元数据
[project]
name = "Fcatbot" # 包名称
version = "2.4.0-alpha" # 版本号
description = "基于ncatbot开发" # 简短描述
authors = [
    { name = "Fish-LP", email = "fish.zh@outlook.com" }, # 作者姓名和邮箱
]
license = { file = "LICENSE" } # 许可证文件路径
readme = "README.md" # 项目说明文件路径
urls = { Homepage = "https://github.com/Fish-LP/FcatBot" } # 项目主页
classifiers = [
    "Development Status :: 3 - Alpha",      # 开发状态
    "Intended Audience :: Developers",      # 目标受众
    "License :: Other/Proprietary License", # 自定义许可协议
    "Programming Language :: Python :: 3",  # 支持的 Python 版本
    # "Programming Language :: Python :: 3.9",  # 不明确是否仍然支持
    # "Programming Language :: Python :: 3.10", # 不明确是否仍然支持
    # "Programming Language :: Python :: 3.11", # 不明确是否仍然支持
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "websockets==10.4",                                  # ws连接
    "packaging~=24.2",                                   # 插件系统
    "tqdm~=4.67.1",                                      # 进度条
    "aiofiles~=24.1.0",                                  # UniversalLoaderError支持异步文件读写
    "schedule~=1.2.2",                                   # 定时任务
    "watchdog~=6.0.0",                                   # 文件监控
    "windows-curses~=2.4.1; platform_system=='Windows'", # Windows平台下的curses支持
    "pyreadline3; platform_system=='Windows'",           # Windows平台下的readline支持
]
optional-dependencies = { "yaml" = [
    "pyyaml ~=6.0.2",
], "toml" = [
    "pytoml ~=0.1.21",
], "ujson" = [
    "ujson~=5.10.0",
] }
requires-python = ">=3.9" # 所需 Python 版本

# Setuptools 特定的配置
[tool.setuptools]
# packages = ["Fcatbot"]      # 自动发现所有包
include-package-data = true # 包含包中的数据文件

# [project.entry-points.console_scripts]
# fcatbot = "Fcatbot.__main__:main" # 命令行启动
